Moved linters into their own directory 🗂️
