+++ src/linters/index.mjs
@@ -1,0 +1,9 @@
+import eslint from './eslint.mjs'
+import stylelint from './stylelint.mjs'
+import markdownlint from './markdownlint.mjs'
+
+export {
+  eslint,
+  markdownlint,
+  stylelint,
+}
